patching up reorg
