+++ trulens_eval/trulens_eval/feedback/provider/endpoint/__init__.py
@@ -1,0 +1,5 @@
+from trulens_eval.feedback.provider.endpoint.endpoint import Endpoint
+from trulens_eval.feedback.provider.endpoint.hugs import HuggingfaceEndpoint
+from trulens_eval.feedback.provider.endpoint.openai import OpenAIEndpoint
+
+__all__ = ['Endpoint', 'HuggingfaceEndpoint', 'OpenAIEndpoint']
