+++ trulens_eval/trulens_eval/feedback/provider/endpoint/endpoint.py
@@ -1,5 +1,4 @@
 import inspect
-import json
 import logging
 from pprint import PrettyPrinter
 from queue import Queue
@@ -7,29 +6,23 @@ from threading import Thread
 from time import sleep
 from types import AsyncGeneratorType
 from types import ModuleType
-from typing import (
-    Any, Awaitable, Callable, Dict, List, Optional, Sequence, Tuple, Type,
-    TypeVar
-)
+from typing import (Any, Awaitable, Dict, Optional, Sequence,
+                    Tuple, Type, TypeVar)
 
-from langchain.callbacks.openai_info import OpenAICallbackHandler
-from langchain.schema import Generation
-from langchain.schema import LLMResult
 import pydantic
 import requests
 
-from trulens_eval.keys import _check_key
-from trulens_eval.keys import get_huggingface_headers
 from trulens_eval.schema import Cost
 from trulens_eval.util import get_first_local_in_call_stack
 from trulens_eval.util import JSON
 from trulens_eval.util import SerialModel
 from trulens_eval.util import SingletonPerName
-from trulens_eval.util import WithClassInfo
+
 from trulens_eval.utils.python import Thunk
-from trulens_eval.utils.text import UNICODE_CHECK
+
 
 logger = logging.getLogger(__name__)
+
 pp = PrettyPrinter()
 
 T = TypeVar("T")
@@ -60,94 +53,6 @@ class EndpointCallback(SerialModel):
 
     def handle_classification(self, response: Any) -> None:
         self.handle(response)
-
-
-class HuggingfaceCallback(EndpointCallback):
-
-    def handle_classification(self, response: requests.Response) -> None:
-        # Huggingface free inference api doesn't seem to have its own library
-        # and the docs say to use `requests`` so that is what we instrument and
-        # process to track api calls.
-
-        super().handle_classification(response)
-
-        if response.ok:
-            self.cost.n_successful_requests += 1
-            content = json.loads(response.text)
-
-            # Huggingface free inference api for classification returns a list
-            # with one element which itself contains scores for each class.
-            self.cost.n_classes += len(content[0])
-
-
-class OpenAICallback(EndpointCallback):
-
-    class Config:
-        arbitrary_types_allowed = True
-
-    # For openai cost tracking, we use the logic from langchain mostly
-    # implemented in the OpenAICallbackHandler class:
-    langchain_handler: OpenAICallbackHandler = pydantic.Field(
-        default_factory=OpenAICallbackHandler, exclude=True
-    )
-
-    chunks: List[Generation] = pydantic.Field(
-        default_factory=list, exclude=True
-    )
-
-    def handle_classification(self, response: Dict) -> None:
-        # OpenAI's moderation API is not text generation and does not return
-        # usage information. Will count those as a classification.
-
-        super().handle_classification(response)
-
-        if "categories" in response:
-            self.cost.n_successful_requests += 1
-            self.cost.n_classes += len(response['categories'])
-
-    def handle_generation_chunk(self, response: Any) -> None:
-        """
-        Called on every streaming chunk from an openai text generation process.
-        """
-
-        # self.langchain_handler.on_llm_new_token() # does nothing
-
-        super().handle_generation_chunk(response=response)
-
-        self.chunks.append(response)
-
-        if response.generation_info['choices'][0]['finish_reason'] == 'stop':
-            llm_result = LLMResult(
-                llm_output=dict(
-                    token_usage=dict(),
-                    model_name=response.generation_info['model']
-                ),
-                generations=[self.chunks]
-            )
-            self.chunks = []
-            self.handle_generation(response=llm_result)
-
-    def handle_generation(self, response: LLMResult) -> None:
-        """
-        Called upon a non-streaming text generation or at the completion of a
-        streamed generation.
-        """
-
-        super().handle_generation(response)
-
-        self.langchain_handler.on_llm_end(response)
-
-        # Copy over the langchain handler fields we also have.
-        for cost_field, langchain_field in [
-            ("cost", "total_cost"), ("n_tokens", "total_tokens"),
-            ("n_successful_requests", "successful_requests"),
-            ("n_prompt_tokens", "prompt_tokens"),
-            ("n_completion_tokens", "completion_tokens")
-        ]:
-            setattr(
-                self.cost, cost_field,
-                getattr(self.langchain_handler, langchain_field)
-            )
 
 
 class Endpoint(SerialModel, SingletonPerName):
@@ -324,6 +229,7 @@ class Endpoint(SerialModel, SingletonPerName):
             obj = getattr(mod, m)
             self._instrument_class(obj, method_name=method_name)
 
+    # TODO: CODEDUP
     @staticmethod
     def track_all_costs(
         thunk: Thunk[T],
@@ -338,6 +244,9 @@ class Endpoint(SerialModel, SingletonPerName):
         endpoints = []
 
         if with_openai:
+            # TODO: DEPS
+            from trulens_eval.feedback.provider.endpoint.openai import OpenAIEndpoint
+
             try:
                 e = OpenAIEndpoint()
                 endpoints.append(e)
@@ -348,6 +257,9 @@ class Endpoint(SerialModel, SingletonPerName):
                 )
 
         if with_hugs:
+            # TODO: DEPS
+            from trulens_eval.feedback.provider.endpoint.hugs import HuggingfaceEndpoint
+
             try:
                 e = HuggingfaceEndpoint()
                 endpoints.append(e)
@@ -359,6 +271,7 @@ class Endpoint(SerialModel, SingletonPerName):
 
         return Endpoint._track_costs(thunk, with_endpoints=endpoints)
 
+    # TODO: CODEDUP
     @staticmethod
     async def atrack_all_costs(
         thunk: Thunk[Awaitable],
@@ -373,6 +286,9 @@ class Endpoint(SerialModel, SingletonPerName):
         endpoints = []
 
         if with_openai:
+            # TODO: DEPS
+            from trulens_eval.feedback.provider.endpoint.openai import OpenAIEndpoint
+
             try:
                 e = OpenAIEndpoint()
                 endpoints.append(e)
@@ -383,6 +299,9 @@ class Endpoint(SerialModel, SingletonPerName):
                 )
 
         if with_hugs:
+            # TODO: DEPS
+            from trulens_eval.feedback.provider.endpoint.hugs import HuggingfaceEndpoint
+
             try:
                 e = HuggingfaceEndpoint()
                 endpoints.append(e)
@@ -835,178 +754,3 @@ class Endpoint(SerialModel, SingletonPerName):
         logger.debug(f"Instrumenting {func.__name__} for {self.name} .")
 
         return w
-
-
-class OpenAIEndpoint(Endpoint, WithClassInfo):
-    """
-    OpenAI endpoint. Instruments "create" methods in openai.* classes.
-    """
-
-    def __new__(cls, *args, **kwargs):
-        return super(Endpoint, cls).__new__(cls, name="openai")
-
-    def handle_wrapped_call(
-        self, func: Callable, bindings: inspect.BoundArguments, response: Any,
-        callback: Optional[EndpointCallback]
-    ) -> None:
-
-        model_name = ""
-        if 'model' in bindings.kwargs:
-            model_name = bindings.kwargs['model']
-
-        results = None
-        if "results" in response:
-            results = response['results']
-
-        counted_something = False
-
-        if 'usage' in response:
-            counted_something = True
-            usage = response['usage']
-
-            llm_res = LLMResult(
-                generations=[[]],
-                llm_output=dict(token_usage=usage, model_name=model_name),
-                run=None
-            )
-
-            self.global_callback.handle_generation(response=llm_res)
-
-            if callback is not None:
-                callback.handle_generation(response=llm_res)
-
-        if 'choices' in response and 'delta' in response['choices'][0]:
-            # Streaming data.
-
-            content = response['choices'][0]['delta'].get('content')
-
-            gen = Generation(text=content or '', generation_info=response)
-            self.global_callback.handle_generation_chunk(gen)
-            if callback is not None:
-                callback.handle_generation_chunk(gen)
-
-            counted_something = True
-
-        if results is not None:
-            for res in results:
-                if "categories" in res:
-                    counted_something = True
-                    self.global_callback.handle_classification(response=res)
-
-                    if callback is not None:
-                        callback.handle_classification(response=res)
-
-        if not counted_something:
-            logger.warning(
-                f"Unregonized openai response format. It did not have usage information nor categories:\n"
-                + pp.pformat(response)
-            )
-
-    def __init__(self, *args, **kwargs):
-        # If any of these keys are in kwargs, copy over its value to the env
-        # variable named as the respective value in this dict. If value is None,
-        # don't copy to env. Regardless of env, set all of these as attributes
-        # to openai.
-
-        # https://learn.microsoft.com/en-us/azure/cognitive-services/openai/how-to/switching-endpoints
-        CONF_CLONE = dict(
-            api_key="OPENAI_API_KEY",
-            organization=None,
-            api_type=None,
-            api_base=None,
-            api_version=None
-        )
-
-        import os
-
-        import openai
-
-        for k, v in CONF_CLONE.items():
-            if k in kwargs:
-                print(f"{UNICODE_CHECK} Setting openai.{k} explicitly.")
-                setattr(openai, k, kwargs[k])
-
-                if v is not None:
-                    print(f"{UNICODE_CHECK} Env. var. {v} set explicitly.")
-                    os.environ[v] = kwargs[k]
-            else:
-                if v is not None:
-                    # If no value were explicitly set, check if the user set up openai
-                    # attributes themselves and if so, copy over the ones we use via
-                    # environment vars, to its respective env var.
-
-                    attr_val = getattr(openai, k)
-                    if attr_val is not None and attr_val != os.environ.get(v):
-                        print(
-                            f"{UNICODE_CHECK} Env. var. {v} set from openai.{k} ."
-                        )
-                        os.environ[v] = attr_val
-
-        # Will fail if key not set:
-        _check_key("OPENAI_API_KEY")
-
-        if hasattr(self, "name"):
-            # Already created with SingletonPerName mechanism
-            return
-
-        kwargs['name'] = "openai"
-        kwargs['callback_class'] = OpenAICallback
-
-        # for WithClassInfo:
-        kwargs['obj'] = self
-
-        super().__init__(*args, **kwargs)
-
-        self._instrument_module_members(openai, "create")
-        self._instrument_module_members(openai, "acreate")
-
-
-class HuggingfaceEndpoint(Endpoint, WithClassInfo):
-    """
-    Huggingface. Instruments the requests.post method for requests to
-    "https://api-inference.huggingface.co".
-    """
-
-    def __new__(cls, *args, **kwargs):
-        return super(Endpoint, cls).__new__(cls, name="huggingface")
-
-    def handle_wrapped_call(
-        self, func: Callable, bindings: inspect.BoundArguments,
-        response: requests.Response, callback: Optional[EndpointCallback]
-    ) -> None:
-        # Call here can only be requests.post .
-
-        if "url" not in bindings.arguments:
-            return
-
-        url = bindings.arguments['url']
-        if not url.startswith("https://api-inference.huggingface.co"):
-            return
-
-        # TODO: Determine whether the request was a classification or some other
-        # type of request. Currently we use huggingface only for classification
-        # in feedback but this can change.
-
-        self.global_callback.handle_classification(response=response)
-
-        if callback is not None:
-            callback.handle_classification(response=response)
-
-    def __init__(self, *args, **kwargs):
-        # Will fail if key not set:
-        _check_key("HUGGINGFACE_API_KEY")
-
-        if hasattr(self, "name"):
-            # Already created with SingletonPerName mechanism
-            return
-
-        kwargs['name'] = "huggingface"
-        kwargs['post_headers'] = get_huggingface_headers()
-        kwargs['callback_class'] = HuggingfaceCallback
-
-        # for WithClassInfo:
-        kwargs['obj'] = self
-
-        super().__init__(*args, **kwargs)
-
-        self._instrument_class(requests, "post")
