+++ trulens_eval/trulens_eval/utils/llama.py
@@ -60,6 +60,7 @@ def component_of_json(json: JSON) -> app.LlamaIndexComponent:
 
 
 class Is:
+    # TODEP
     """
     TODO: DEPRECATE: Replacing with component view types.
