+++ trulens_eval/trulens_eval/utils/text.py
@@ -2,6 +2,10 @@
 Utilities for user-facing text generation.
 """
 
+import logging
+
+logger = logging.getLogger(__name__)
+
 UNICODE_STOP = "🛑"
 UNICODE_CHECK = "✅"
 UNICODE_YIELD = "⚡"
